feat(agent): migrate truncation limits to reactive LLM events (Phase 3)
Extends the events system (Phase 1: auto-compact/fold; Phase 2:
round-count guards) to five byte-length truncation constants:

- maxToolResultBytes -> tool-result-truncate (truncate)
- toolErrorMaxLen -> tool-error-truncate (truncate)
- maxForceAnswerDigestBytes -> force-answer-digest (truncate)
- toolLoopMessageBudget -> history-window-trim (drop_oldest_roundtrip)
- maxFileReadBytes -> file-read-limit (reject)

Adds EventTrigger.Bytes and effectiveBytes(name, fallback) -- unlike
effectiveRounds's universal floor of 1, byte thresholds span 500 to 1MB
across this cluster so the safety fallback is per-call-site. All five
migrated constants became zero-arg functions since Go can't source a
const from a runtime lookup; call sites needing int64 comparisons got
an explicit conversion where they used to rely on untyped-constant
auto-conversion. maxFileReadBytes had the widest blast radius (9 call
sites across 4 files) but was a pure mechanical substitution.

+++ pkg/agent/builtin_resource_tools.go
@@ -445,7 +445,7 @@ func annotateSearchMatches(results []map[string]interface{}, query string) {
 			continue
 		}
 		info, statErr := AppFS.Stat(path)
-		if statErr != nil || info.IsDir() || info.Size() > maxFileReadBytes {
+		if statErr != nil || info.IsDir() || info.Size() > int64(maxFileReadBytes()) {
 			continue
 		}
 		data, readErr := afero.ReadFile(AppFS, path)
